+++ app/pages/profile/profile.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router';
-import './profile.less';
+import './profile.scss';
 
 export default class Profile extends React.Component {
   constructor(props) {
@@ -16,7 +16,6 @@ export default class Profile extends React.Component {
             <ul className="nav nav-pills nav-stacked">
               <li role="presentation"><Link to={'/home'}>Home</Link></li>
               <li role="presentation" className="active"><Link to={'/profile'}>Profile</Link></li>
-              <li role="presentation"><Link to={'/register'}>Register</Link></li>
               <li role="presentation"><a href="#">Go Back</a></li>
             </ul>
           </div>