+++ app/utils/ie-compatible.js
@@ -2,7 +2,6 @@ import 'html5shiv';
 import 'es5-shim';
 import 'es5-shim/es5-sham';
 import 'console-polyfill';
-// import 'babel-polyfill';
-import 'fetch-ie8';
+import 'babel-polyfill';
 import 'core-js/fn/object/assign';
 import '../../lib/respond/respond.modified.min';